fixes includes and class definitions for non-unified build
