+++ backends/graphs/parser.cpp
@@ -20,6 +20,8 @@ limitations under the License.
 
 #include "graphs.h"
 
+#include "frontends/common/resolveReferences/referenceMap.h"
+#include "frontends/p4/typeMap.h"
 #include "lib/log.h"
 #include "lib/nullstream.h"
 #include "lib/path.h"
